// types/enums.ts
export type SkinLevel = "default" | "epic" | "legend";
export type Difficulty = "mudah" | "sedang" | "sulit";
export type RoomStatus = "open" | "playing" | "processing" | "finished";
export type RoomVisibility = "public" | "private";
export type CategoryType =
  | "bahasaindonesia"
  | "bahasainggris"
  | "biologi"
  | "pancasila"
  | "pemrograman"
  | "sejarah";